convergence-data-pipeline/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Cloud Platform
google-cloud-bigquery==3.14.1
google-cloud-secret-manager==2.18.1
google-cloud-logging==3.9.0
google-cloud-trace==1.12.0

# Async Workers
celery[redis]==5.3.4
redis==5.0.1

# Data Processing
polars==0.20.3
pyarrow==15.0.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Quality
great-expectations==0.18.8

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-gcp-trace==1.6.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-requests==0.43b0
python-json-logger==2.0.7

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25

# Utilities
tenacity==8.2.3
pendulum==3.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
mypy==1.8.0
